# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Photography Project Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CreateClientBody {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateClientBody {
  name?: string;
  email?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  inquiry: "inquiry",
  booked: "booked",
  in_progress: "in_progress",
  editing: "editing",
  delivered: "delivered",
  archived: "archived",
} as const;

export type ProjectProjectType =
  (typeof ProjectProjectType)[keyof typeof ProjectProjectType];

export const ProjectProjectType = {
  wedding: "wedding",
  portrait: "portrait",
  commercial: "commercial",
  event: "event",
  product: "product",
  fashion: "fashion",
  real_estate: "real_estate",
  other: "other",
} as const;

export interface Project {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  projectType: ProjectProjectType;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  sessionDate?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateProjectBodyStatus =
  (typeof CreateProjectBodyStatus)[keyof typeof CreateProjectBodyStatus];

export const CreateProjectBodyStatus = {
  inquiry: "inquiry",
  booked: "booked",
  in_progress: "in_progress",
  editing: "editing",
  delivered: "delivered",
  archived: "archived",
} as const;

export type CreateProjectBodyProjectType =
  (typeof CreateProjectBodyProjectType)[keyof typeof CreateProjectBodyProjectType];

export const CreateProjectBodyProjectType = {
  wedding: "wedding",
  portrait: "portrait",
  commercial: "commercial",
  event: "event",
  product: "product",
  fashion: "fashion",
  real_estate: "real_estate",
  other: "other",
} as const;

export interface CreateProjectBody {
  title: string;
  /** @nullable */
  description?: string | null;
  status: CreateProjectBodyStatus;
  projectType: CreateProjectBodyProjectType;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  sessionDate?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type UpdateProjectBodyStatus =
  (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus];

export const UpdateProjectBodyStatus = {
  inquiry: "inquiry",
  booked: "booked",
  in_progress: "in_progress",
  editing: "editing",
  delivered: "delivered",
  archived: "archived",
} as const;

export type UpdateProjectBodyProjectType =
  (typeof UpdateProjectBodyProjectType)[keyof typeof UpdateProjectBodyProjectType];

export const UpdateProjectBodyProjectType = {
  wedding: "wedding",
  portrait: "portrait",
  commercial: "commercial",
  event: "event",
  product: "product",
  fashion: "fashion",
  real_estate: "real_estate",
  other: "other",
} as const;

export interface UpdateProjectBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: UpdateProjectBodyStatus;
  projectType?: UpdateProjectBodyProjectType;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  sessionDate?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type ListProjectsParams = {
  status?: ListProjectsStatus;
  clientId?: number;
};

export type ListProjectsStatus =
  (typeof ListProjectsStatus)[keyof typeof ListProjectsStatus];

export const ListProjectsStatus = {
  inquiry: "inquiry",
  booked: "booked",
  in_progress: "in_progress",
  editing: "editing",
  delivered: "delivered",
  archived: "archived",
} as const;