# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Photography Project Management API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all clients
 */
export const ListClientsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.date(),
});
export const ListClientsResponse = zod.array(ListClientsResponseItem);

/**
 * @summary Create a new client
 */
export const CreateClientBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  phone: zod.string().nullish(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get a client by ID
 */
export const GetClientParams = zod.object({
  id: zod.coerce.number(),
});

export const GetClientResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.date(),
});

/**
 * @summary Update a client
 */
export const UpdateClientParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateClientBody = zod.object({
  name: zod.string().optional(),
  email: zod.string().optional(),
  phone: zod.string().nullish(),
  notes: zod.string().nullish(),
});

export const UpdateClientResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.date(),
});

/**
 * @summary Delete a client
 */
export const DeleteClientParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all projects
 */
export const ListProjectsQueryParams = zod.object({
  status: zod
    .enum([
      "inquiry",
      "booked",
      "in_progress",
      "editing",
      "delivered",
      "archived",
    ])
    .optional(),
  clientId: zod.coerce.number().optional(),
});

export const ListProjectsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum([
    "inquiry",
    "booked",
    "in_progress",
    "editing",
    "delivered",
    "archived",
  ]),
  projectType: zod.enum([
    "wedding",
    "portrait",
    "commercial",
    "event",
    "product",
    "fashion",
    "real_estate",
    "other",
  ]),
  clientId: zod.number().nullish(),
  clientName: zod.string().nullish(),
  sessionDate: zod.date().nullish(),
  location: zod.string().nullish(),
  price: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListProjectsResponse = zod.array(ListProjectsResponseItem);

/**
 * @summary Create a new project
 */
export const CreateProjectBody = zod.object({
  title: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum([
    "inquiry",
    "booked",
    "in_progress",
    "editing",
    "delivered",
    "archived",
  ]),
  projectType: zod.enum([
    "wedding",
    "portrait",
    "commercial",
    "event",
    "product",
    "fashion",
    "real_estate",
    "other",
  ]),
  clientId: zod.number().nullish(),
  sessionDate: zod.date().nullish(),
  location: zod.string().nullish(),
  price: zod.string().nullish(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get a project by ID
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum([
    "inquiry",
    "booked",
    "in_progress",
    "editing",
    "delivered",
    "archived",
  ]),
  projectType: zod.enum([
    "wedding",
    "portrait",
    "commercial",
    "event",
    "product",
    "fashion",
    "real_estate",
    "other",
  ]),
  clientId: zod.number().nullish(),
  clientName: zod.string().nullish(),
  sessionDate: zod.date().nullish(),
  location: zod.string().nullish(),
  price: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update a project
 */
export const UpdateProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProjectBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().nullish(),
  status: zod
    .enum([
      "inquiry",
      "booked",
      "in_progress",
      "editing",
      "delivered",
      "archived",
    ])
    .optional(),
  projectType: zod
    .enum([
      "wedding",
      "portrait",
      "commercial",
      "event",
      "product",
      "fashion",
      "real_estate",
      "other",
    ])
    .optional(),
  clientId: zod.number().nullish(),
  sessionDate: zod.date().nullish(),
  location: zod.string().nullish(),
  price: zod.string().nullish(),
  notes: zod.string().nullish(),
});

export const UpdateProjectResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum([
    "inquiry",
    "booked",
    "in_progress",
    "editing",
    "delivered",
    "archived",
  ]),
  projectType: zod.enum([
    "wedding",
    "portrait",
    "commercial",
    "event",
    "product",
    "fashion",
    "real_estate",
    "other",
  ]),
  clientId: zod.number().nullish(),
  clientName: zod.string().nullish(),
  sessionDate: zod.date().nullish(),
  location: zod.string().nullish(),
  price: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a project
 */
export const DeleteProjectParams = zod.object({
  id: zod.coerce.number(),
});