# lib/api-zod/src/generated/types/createClientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Photography Project Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateClientBody {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  notes?: string | null;
}