# lib/api-zod/src/generated/types/createProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Photography Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProjectBodyProjectType } from "./createProjectBodyProjectType";
import type { CreateProjectBodyStatus } from "./createProjectBodyStatus";

export interface CreateProjectBody {
  title: string;
  /** @nullable */
  description?: string | null;
  status: CreateProjectBodyStatus;
  projectType: CreateProjectBodyProjectType;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  sessionDate?: Date | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  notes?: string | null;
}