# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Photography Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectProjectType } from "./projectProjectType";
import type { ProjectStatus } from "./projectStatus";

export interface Project {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  projectType: ProjectProjectType;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  sessionDate?: Date | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}