# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Photography Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyProjectType } from "./updateProjectBodyProjectType";
import type { UpdateProjectBodyStatus } from "./updateProjectBodyStatus";

export interface UpdateProjectBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: UpdateProjectBodyStatus;
  projectType?: UpdateProjectBodyProjectType;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  sessionDate?: Date | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  notes?: string | null;
}